requirements.txt
annotated-doc
annotated-types
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click
fastapi
h11
ibm-cos-sdk
ibm-cos-sdk-core
ibm-cos-sdk-s3transfer
ibm_db
idna
importlib_metadata
jmespath
lomond
numpy
packaging
pandas==2.1.4
ibm_watson_machine_learning==1.0.368
pydantic
pydantic_core
python-dateutil
python-dotenv
pytz
requests
six
sniffio
starlette
tabulate
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
zipp
